fix(BE): 메서드 중복 문제 해결
- get_current_user 메서드 병합 및 정리
