+++ app/backend/src/main/middleware/auth.py
@@ -1,7 +1,8 @@
 from fastapi import HTTPException, Request
 from fastapi.security import HTTPBearer
 
-from src.main.core.auth.dependencies import get_current_user
+from src.main.domains.user.dependencies import get_redis_token_manager
+from src.main.core.auth.dependencies import get_current_user, oauth2_scheme
 from src.main.db.deps import get_db
 
 
@@ -11,15 +12,19 @@ async def auth_middleware(request: Request, call_next):
     try:
         # DB 세션 획득
         db = await get_db().__anext__()
-        token = request.headers.get("Authorization")
-        if token and token.startswith("Bearer "):
-            token = token.split(" ")[1]
-            user = await get_current_user(token, db)
-            request.state.user = user
+        token_manager = get_redis_token_manager()
+        token = await oauth2_scheme(request)
+        if token:
+            try:
+                user = await get_current_user(token, db, token_manager)
+                request.state.user = user
+            except HTTPException:
+                # 인증 실패 시 처리
+                request.state.user = None
         else:
             request.state.user = None
-    except HTTPException:
-        # 인증 실패 시 처리
+    except Exception as e:
+        print(f"Auth middleware error: {str(e)}")
         request.state.user = None
 
     response = await call_next(request)
